[skrifa] some module reorg and cleanup
This is work I wanted to do before publishing but didn't get around to it. Looks like a lot of churn here, but no functional changes.

* removes the unnecessary `meta` module and moves submodules to the crate root
* moves some types into new modules (`FontRef` and `UniqueId` move to a new `font` module)
* adds a `prelude` for help with importing common types (`FontRef`, `GlyphId`, `Size`, etc).

This one is biggish, but If there are no issues, just merge!
