--- skrifa/src/meta/mod.rs
@@ -1,9 +1,0 @@
-//! Accessors for various types of font metadata.
-
-pub mod charmap;
-pub mod metrics;
-pub mod strings;
-
-mod provider;
-
-pub use provider::MetadataProvider;
